+++ tests/docsrc/doc_metarepo.cpp
@@ -249,9 +249,9 @@ The first templated form is similar to `getDerives(metapp::getMetaType<remove al
 
 ```c++
 template <typename Class>
-void * castToBase(void * instance, const size_t baseIndex) const;
+void * castToBase(void * instance, const int baseIndex) const;
 
-void * castToBase(void * instance, const MetaType * classMetaType, const size_t baseIndex) const;
+void * castToBase(void * instance, const MetaType * classMetaType, const int baseIndex) const;
 ```
 
 Casts `instance` of `Class`, or of `classMetaType`, to its base class at `baseIndex` in the base list returned by `getBases`, then returns the casted pointer.  
@@ -261,9 +261,9 @@ The first templated form is similar to `castToBase(instance, metapp::getMetaType
 
 ```c++
 template <typename Class>
-void * castToDerived(void * instance, const size_t derivedIndex) const;
+void * castToDerived(void * instance, const int derivedIndex) const;
 
-void * castToDerived(void * instance, const MetaType * classMetaType, const size_t derivedIndex) const;
+void * castToDerived(void * instance, const MetaType * classMetaType, const int derivedIndex) const;
 ```
 
 Casts `instance` of `Class`, or of `classMetaType`, to its derived class at `derivedIndex` in the derived list returned by `getDerives`, then returns the casted pointer.
